backend/requirements.txt
django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
model-bakery==1.20.5
openai==1.101.0
psycopg[binary]==3.2.1
uvicorn[standard]==0.32.0
